packetbuffer: fix logic for FreePacket()
m_next_empty_packet_key is a random number that is incremented.  In FreePacket(),
the upper 32 bits are tested for equivalence.

However, the incrementing of m_next_empty_packet_key could overflow the lower 32 bits,
changing the upper 32.  This would cause the if statement to be false and the packet to
not be marked as empty.

Starting with m_next_empty_packet_key with the lower 32 bits cleared, prevents this error
by allowing 2^32 packets before overflowing, far more than can be reasonably expected.