Add recordedid and pathname to recording info window
Add recordedid and pathname to the recorded program info window.
In the Watch Recordings screen, selecting a recording and then press "i" will
give a popup window with information about the recording.
Pressing "i" a second time add more details.
This information is added to the "second i" window;

Fixes #464