[webapp] First pass at the TV Guide
This provides the most basic TV Guide possible,
showing a 2hr window from the current time for
all channels.

Copies tv/images to images/tv, the former will
go away with the removal of QtScript based code
and the latter is the "right place" for images
to live.

TODO:
- Add support for a time / date picker to navigate
through the guide.
- Add lazy loading and pagination to improve load
times for large channel sets.
- Add the ability to interact with the guide
  * drill down to program details
  * schedule recordings
- Do some proper styling. This styling is primarily
copied from the old web frontend.