Work around an FF/Rew issue when using mediacodec/surface.
FF/Rew works by repeatedly decoding a keyframe and then seeking to a
new keyframe some distance away.  This usually results in several
keyframes in the videobuffer queue waiting to be displayed.
Unfortunately, when using mediacodec with surface rendering, Android
appears to not be able to render a frame after a seek has occurred.
The result is that the video appears frozen with the last, rendered
frame while the ff/rew continues.

This commit works around the issue by enabling a new, "render one"
mode for ff/rew when using mediacodec/surface.  In this mode, after
decoding a keyframe and queuing it for display, the decoder waits for
the frame to be rendered before seeking to the next keyframe.  In
effect, it synchronizes the decoder and display loops to handling one
frame at a time.