webapp: add support for translations
The webapp uses ngx-translate to load and handle switching languages.

There are some initial *.json files for all the languages we support in MythTV.

There are several editors available for translators to update the language json
files (https://github.com/ngx-translate/core#editors) the best of which is
BabelEdit but it is a paid for app but because the translation files are just
plain JSON files they can be edited with any text editor.

For developers who use VSCode there are several plugins available but the
"Lingua - Managing nxg-translations" plugin seems to be one of the better ones.