More room for "Please Wait..." on playback screen
After a recording is selected for playback and before the playback
actually starts the message "Please Wait..." is presented.
In theme MythCenter-wide the width of the message box is just big enough.
Make the box a bit bigger so that also slightly longer texts, such as
the Dutch translation, can be presented here.