More room for "Parental Level" on Videos screen
In the Videos screen there is a "Parental Level" message presented
at the right/bottom corner.
In theme MythCenter-wide the width of the message box is just big enough.
Make the box a bit bigger so that also slightly longer texts, such as
the Dutch translation, can be presented here.