Fix "failed to find recorded entry for 0" warnings.
Add a couple of checks to the data flow when updating a ProgramInfo
data structure. First, before sending a notification that a
ProgramInfo has changed ensure that the data structure actually
represents a recording (not a video, DVD, Bluray, or stream).  Second,
when receiving a MASTER_UPDATE_REC_INFO message verify that the
recording id number isn't zero.

Fixes #483.