use FileSystemInfo::Consolidate()
in MainServer::BackendQueryDiskSpace()

The only difference with FileSystemInfo::Consolidate() was the use of
GetfsID() instead of starting from 0 every time.  I'm not sure why it
wasn't removed when the class was created in 65e877ae9c2e2bbbfeeb00740aebf511bf588c84.

The only change is that the m_fsid may not be the same for the same
hostname and path on subsequent calls to BackendQueryDiskSpace(),
which would be a wierd thing the rely on especially since this was the only
function that did that.