fixup QtEndian template use (s390x)
The error was:
/usr/bin/ld: ../../libs/libmyth/libmyth-32.so: undefined reference to `unsigned long qbswap<unsigned long>(unsigned long)'

This must be from eldutils.cpp: uint64_t is unsigned long, but Qt must
be using quint64 = unsigned long long, which causes the link error.

To fix this, follow the Qt documentation and only use the allowed
types, i.e. quint16, qint16, quint32, qint32, quint64, or qint64.