configure: set -mtune=native for --cpu=host
For GCC:
x86: "Using -mtune=native produces code optimized for the local
machine under the constraints of the selected instruction set."

AArch64: "Additionally on native AArch64 GNU/Linux systems the value
‘native’ tunes performance to the host system. This option has no
effect if the compiler is unable to recognize the processor of the
host system."

Similar for other architectures.

Note: this is equivalent to -mcpu=native for architectures that support
that option; notably, for x86, -mcpu is a deprecated synonym of -mtune.
However, on x86, -march implies -mtune in this and most cases.

For clang: from its command line reference for -mtune=<arg>:
"Only supported on X86 and RISC-V. Otherwise accepted for compatibility
with GCC."