Remove extraneous ? character in Sat>IP OPTIONS message
The extraneous ? character occurs when the QUrl query is initialized to an empty string "".
Fixed by replacing the "" by a QString() in the OPTIONS and the PLAY query composition.
Thanks to Mike Bibbings for reporting and fixing this for the TEARDOWN command,
see https://code.mythtv.org/trac/ticket/13121#comment:20

Refs #13121