Allow embedded video in EPG when near program end in live TV.
When near the end of a recorded program or real time in live TV or
watching while recording and the user tries to launch the EPG or other
guide screen, the TV code always puases playback.  This disables the
use of the guide screen that supports embedded video.  This change
disables the near end check for live TV as playback can and will
continue with the following program on the current channel.