MythDate: replace formatTime() with formatDuration()
QTime will only print up to 23 hours, 59 minutes, 59 seconds, and 999
milliseconds.  This implementation has none of those restrictions.

Fixes https://github.com/MythTV/mythtv/issues/520

However, some of the output is changed: in particular, zero padding is
now the default, but can be disabled if desired.

As far as I can tell, this is only used for human readable output.