Remove reference from libmythbase to libmyth.
The mythpluginapi.h file should not be reaching outside of its
directory to pull in a file from libmyth. (Libmythbase isn't supposed
to have any dependencies.) Move the only usage of the MPLUGIN_API
define to make this out-of-directory include unnecessary. Requires
changing the plugins to consistently include mythbaseexp.h (instead of
a mix of mythbaseexp.h and mythexp.h), and changing from the MPUBLIC
symbol to the MBASE_PUBLIC symbol.