Base watch list ordering on series most recently warched or recorded
(Also fixes the problme of air date taking priority over series and
episode)

Before this change, the watch list used a complex ordering that was
designed to help users keep up to date with many series simultaneosly,
ensuring that they didn't get behind with some. It's cleverly designed,
but seems very specific to the author's particular viewing habits. It
doesn't account well for shows that you really don't want to watch just
now, but you think you might one day. Nor does it account for a wish to
wait for a whole series before binging.

This commit simplifies the design, taking the view that mostly the user
will want to continue with series that they've recently viewed, or to
which a new recording has recently been added. The new ordering gives the
user more control: if you wish to make a series appear higher in the watch
list then just watch a few seconds of an episode; if you wish to make a
series appear lower then don't watch any episodes from it and it will
get pushed out by the series you are watching.