reduce duplicate filenames (main.cpp and commandlineparser.*) (#533)
This is intended to reduce confusion in the Doxygen docs and also any errors while compiling.

In the programs directory, this makes the file sort with the project
file and the compiled binary.  In the plugins directory, this makes
the file sort with the project file, but not the compiled binary,
which is prefixed by lib.

main_helpers was split off in 064931c34f3b9176970f93f1929b7e6091254b58 ,
so I have also renamed those to mythbackend_main_helpers.

The old mythutil.cpp (one function) and main.cpp files were merged into
a single mythutil.cpp file.