Update preview picture after record playback exit
When exiting a record playback, the preview thumbnail picture that is shown
at the right/bottom corner of the MythCenter-wide theme, is updated to show
a still image of the recording at the last played position.
Since commit 241d26d42ed625b069d9f8412030679429b14ca5, the 'Add percentage
finished to "View Recordings" and "Upcoming Recordings"', issue #331, the
preview picture is sometimes updated and sometimes not.
This is now fixed by using a new field m_previewUpdate to store a copy of
m_bookmarkUpdate when preview picture generation has been requested.
This field is then used to compare against m_bookmarkUpdate
to flag if the preview picture has to be updated.

Fixes #547