Animation update interval lower bound
Add a lower bound of 10ms for the animation update interval.
The animation is used in the MythBuntu theme and there it can cause a stall of
the GUI when the "Manage Recordings" / "Previously Recorded" screen is selected.
This seems to happen only when there is a small number of recordings.
The animation update draw interval was previously fixed at 16 milliseconds
as computed by the function GetDrawInterval.
This function has been removed in commit 34dfa31bc717628bc20a88dd396d54a3ef56b824
and the update interval has since been dynamically computed with an upper
bound of 50ms but without a lower bound.
There is now a lower bound of 10ms added so that the computed interval time
is clamped between 10ms and 50ms.