EIT start delay
Compute the second component of the EIT start delay with the highest
input number instead of with the number of inputs.
The input number can be higher than the number of inputs because inputs (capture cards)
can be deleted and because new inputs always get a higher number.
This then can lead to a EIT start delay that is larger than intended.