tidy: Fix "multiplication implicit widening" in mythcommflag.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

Fix this code to use int64_t for everything, instead of calculating in
an int64_t and truncating to an int, then converting back to int64_t
for the next calculation, and then truncating again, etc.