tidy: Fix "multiplication implicit widening" in mythshutdown/welcome.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

In these cases, the fix is simply to force the "multiplication by -1"
to happen as a 64 bit number.