tidy: Fix "multiplication implicit widening" in mythzoneminder.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

This set of warnings are false positives as the compiler is generating
constant values that are well within the range of the type. The fix is
simply to force the compiler to perform the multiplication to happen
as a 64bit number.