tidy: Fix "multiplication implicit widening" in mythcdrom.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

Fix by forcing the compiler to perform the multiplication as a 64bit
number.