Fix problem with GetStatus API call while playing video.
This bug was introduced while fixing a problem where all the network
control "play" commands were duplicated by MythTV.  Most of the "play"
commands are idempotent, so the only place the duplication was visible
was the "play speed pause" command.

Fixing that problem using a different method restores the
functionality of the GetStatus API call.

Fixes #545.