tidy: Fix "misplaced widening cast" warnings.
This check will warn when there is a cast of a calculation result to a
bigger type. If the intention of the cast is to avoid loss of
precision then the cast is misplaced, and there can be loss of
precision. Otherwise the cast is ineffective.

Fix these warning by casting the first term of the multiplication to
the bigger type, instead of casting the result of the multiplication.