DiSEqC Unicable userband range check
Change the Unicable userband ID range check limit from 8 to 7.
When the Userband ID is out of range this is now treated as an error and not as a warning.
The userband channels are often numbered from 1 upwards but MythTV starts numbering at 0.
The userband number is copied into a 3-bit DiSEqC command field so the maximum is really 7.
Also a few minor cosmetic changes.