Convert int->ssize_t in galleryviews.
This variable is set from a call to a size() function, and its only
use is when it is compared to the result of a size() function.  The
type is ssize_t instead of size_t because the initial call to create
the object uses a value of "xxx.size() - 1" without testing if that
expression yields -1.