tidy: Fix a few "defines a destructor but not..." warnings.
The clang-tidy "special member functions" check looks for violations
of the "rule of five".  This states that if you define any of a
destructor, a copy constructor, a copy assignment, a move constructor,
or a move assignment function, then you need to define all of them.
Non compiler generated default versions of these functions are
generally only needed when managing a dynamically allocated resource,
and they all need to be implemented to properly manage that dynamic
resource.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/special-member-functions.html
https://cpppatterns.com/patterns/rule-of-five.html