MythEDID: fix out of bounds memory access
found by valgrind while investigating segmentation faults that
occured randomly on startup of mythfrontend.

There are only 4 descriptors in the EDID data structure.
By incorrectly looping 5 times in MythEDID::ParseBaseBlock(),
an offset of 126 is passed to ParseDisplayDescriptor() or
ParseDetailedTimingDescriptor() which will then read past
the end of the 128 byte EDID data structure.

This appears to have always been wrong since it was added in
53d170221983764ae108939cbdb0bcbfc48b6114 .