Remove quotes in filelyrics.py argument values
The string quotes around each argument value in the command to start filelyrics.py causes a failure in reading the .lrc file.
This is a regression; behavior of fixes/31 is correct but fixes/32 and master is broken.
These quotes have been added as part of commit b50234ef0229f1c08d32149fcb7937b15b75a603.

Refs #590