tidy: Convert exit code macros to enums.
This requires tweaking the test code to explicitly convert the enum
values to "int" since the Qt metatype code doesn't recognize the enum.
This could be solved by using an "enum class" instead of a plain enum,
but that would cause churn as every reference would need to be
updated (à la RecStatus::Type).  A project for another day.