Remove initializers TSPacket and TSHeader
The initialization was added in commit 8e70b50643611fac80521e6c638aef67dbfabc41 on Jan 10, 2020
to remove warnings from a static code analyzer.
However, the initialization is not needed at all, see comments on lines 52-53 and 61-62, added in
commit 6a4845b04d0f52f9189d08b416c2af2047329db5 on May 1, 2008 and on line 207, added in
commit 4ca584ae1bce9a3a6ddd4e2df27b9a4930c85010 on Jun 21, 2008 by the original author of this code.
Superfluous initializations are not harmful but in this case it can cost a significant amount of processing resources.
For one full transport stream this can be 4 to 5 Mbyte/second that is needlessly initialized.