CardUtil InputGroups: unlink temporary cardid 0 after a real card is linked.
When CardUtil::CreateInputGroup is called, it creates a temporary entry with
a cardid of zero (0). That is no longer needed once a real cardid is linked
to the group to remove it at that time.

videosource.cpp does a general cleanup by calling
CardUtil::UnlinkInputGroup(0,0); which I have left for now, but could be
removed now that it is done automatically whenever a input group link is created.

Yes, doing the cleanup on every link is slightly more inefficient, but the
code flow is cleaner and easier to follow without missing the cleanup step.