cppcheck: Fix "use initialization list" warnings.
Move member variable assignment from the constructor body into the
function initialization list.  In the cases where the function body is
now empty, move the initializer list into the header file.

In one case, a redundant initializer was removed. In another, and
argument was converted to const&. In another, std::move was used with
the argument. These changes were based upon subsequent runs of
clang-tidy.