cppcheck: Fix "use initialization list" warnings (2).
Move member variable assignment from the constructor body into the
function initialization list.

For some reason this change causes the clang-tidy NewDeleteLeaks check
to now determine that some (but not all) of the MusicGenericTree nodes
allocated in PlaylistEditorView::createRootNode are potential memory
leaks.  A pointer to the first node is saved in m_rootNode, and all of
the other nodes are attached as children of that first node (inside
the constructor) so there is no memory leak.  Disable the
NewDeleteLeaks check for that specific function.