Option to exit playback and clear last played position
Add the "Clear last played position and exit" choice to the "Action on playback exit"
configuration in mythfrontend/Setup/Video/Playback/General Playback.
Add the "Exit Without Saving" option to the playback exit menu. This menu is shown when in the
"Action on playback exit" configuration in mythfrontend/Setup/Video/Playback/General Playback
an option for a prompt on playback exit is selected.
In fixes/31 the bookmark representing the last_played_position is only updated
on playback exit when so configured.
In fixes/32 the last_played_position is continuously updated so there is no easy way to do
a real "Exit Without Saving" because the last_played_position has been saved already.
Instead, the "Exit Without Saving" and "Clear last played position and exit" actions
are implemented by clearing the last played position on playback exit.
This insures that the next playback starts at the beginning of the recording or
at the bookmark, when there is a bookmark present.
This is the same as the behavior of fixes/31 when so configured.

The difference with the previous version of this feature in commit c46e186c3b48c07e87331ece16e0858d2d9d4470
is that the code used for choice "Clear last played position and exit" had the same numerical
value as the choice "Save position and exit" in fixes/31. This caused a change in behavior when upgrading
from fixes/31 to fixes/32.
This is now avoided by using a numerical value that has not previously been used in this context.

Refs #543

(cherry picked from commit a03fbcac8789f21c638f58907486e5344ece2975)