FileLogger: convert to C++ I/O
The POSIX write() system call may truncate the output.  The proper
way to use it is in a loop.

Instead, use the high level C++ STL I/O.

The Boost library would be needed to keep using the open() and close()
system calls, i.e. to wrap a FILE* with an iostream.

I don't see a way via the C++ STL ofstream to set the Unix file
permissions.  However, it seems to default to the previously
specified 0664.

I'm not sure how to test that FileLogger::reopen() actually works
with a log roller, but it should be equivalent.