MythMediaMonitor: Replace UDisks with UDisks2
UDisks is long time deprecated and not available in modern Linux
distributions anymore.
It was replaced by UDisks2 which is now provided by the storaged.org
project and linked from freedesktop.org.
See https://udisks.freedesktop.org/ and
http://storaged.org/doc/udisks2-api/latest/ for details.

When monitoring is enabled in the frontend:
Setup -> General ->  Monitor Devices
a 'MythMediaDevice' is added for every removeable device.

Refs https://github.com/MythTV/mythtv/issues/605