Fix fileinfo.xml generation in mytharchivehelper
Plugin mytharchive:
When ffmpeg cannot determine the codec of the provided stream,
the whole buffer provided by `avcodec_string` is taken in the function
`getFileInfo` to create an xml entry. Cleanup this buffer by removing
'\0' characters. Otherwise, the return codec in the xml sequence will
hold 256 zeros and mythburn.py fails to read this xml file.

Refs #609