Housekeeper: allows initial INSERTs to the housekeeping table
Fresh DBs have no entries in housekeeping. INSERTs weren't
being made because the test that decided to do it always
failed and caused an UPDATE. That did nothing because a
matching tag wasn't found.

Now, always do an UPDATE and if the no rows were affected,
then do an INSERT.