Add bounds checking to ATSC table parsing.
When parsing any of the variable length portions of any of the ATSC
PSIP tables, ensure that the parsing never reads past the end of the
packet.  Knowing the number of entries in the table up front also
allows for preallocating all the space needed for the m_ptr array up
front, instead of std::vector making multiple allocation requests as
the vector size grows.  Add error counters that are printed when a
recording is stopped.