bytereader.cpp: tidy fix bugprone-branch-clone
The first two conditions can be tested in either order, so switch
the order to not duplicate the branch.

Compare `< 1` instead of `== 0` to hint to the compiler to emit only one
test instruction against 1 instead of a test against 0 and then a test
against 1.  Since we are scanning memory as uint8_t, they are evquivalent.

Signed-off-by: David Hampton <mythtv@love2code.net>