mpegts-mythtv.c: reuse existing streams for MHEG
This should emulate the prior behavior before harmonization of
removing and adding new streams.

My sample now only has two MHEG streams, one for each PID,
as it used to.

In the process, inline new_section_av_stream to ensure the values
are always set correctly.

I don't know if this is the correct way to reuse the streams,
but it appears to work correctly.