Channel Group per Video Source for Program Guide
Automatically add a Channel Group for each video source.
This group contains the visible channels of only that video source.
This makes it easier to use the Program Guide in large systems
with many video sources and with overlapping channel numbers.
In addition to this there is now also a Channel Group "Priority"
that contains the channels of all video sources that have
the "recpriority" set to a value bigger than 0.
The existing functionality of the "Favorites" channel group
and the capability to add more manually maintained channel groups
is not changed.

Refs #616