Fix warning message logic in DVBSignalMonitor
The status value mok is passed to the lambda functions log_message and
update_rmflags via capture. This however captures the value at
the line where the lambda function is defined; it does not copy
the actual value of variable mok resulting in erroneous error messages.
This is now fixed by adding a & to the capture variable, specifying
capture by reference which means that the actual variable value is used.
This bug is introduced in commit 7c6cee5b031e1eafe609b0740bb7b930ad7b7f03
as part of the continuous fixing of clazy/tidy warnings.