Add support for the DVB "Time Offset Table".
The first part of this table is identical to the DVB "Time and Date
Table".  This table also includes descriptors to "describe country
specific dynamic changes of the local time offset relative to UTC". It
also has a CRC, where the TDT doesn't.  The only valid descriptor for
this table is the "Local time offset descriptor".

1) Decode the "Time Offset Table".

2) Update the existing "Local time offset descriptor" to know that the
values are BCD encoded. (See DVB A038 Section 6.2.20.)