Allow some channel, string settings to be cleared.
ChannelUtil::UpdateChannel() incorrectly prevents some channel
seetings from being cleared/set to empty strings by checking for
isEmpty().  This primarily affects the services API.  Fix the problem
by checking isNull() instead.  When no value is specified via the API,
a null (not NUL) string is passed in and no change is applied.  When
an empty string is specified via the API, however, a non-null (but
empty) string is passed in and the value is correctly cleared.