Fix inefficiency in ChannelUtil::GetChannelsInternal()
The recent addition of supprt for automatic, channel groups and
addition of new channels to my own videosources exposed a huge
inefficiency in ChannelUtil::GetChannelsInternal().  Fix it by
refactoring the query to not create so many rows that need to then be
grouped.

Also fix a bug where the inputIds and groupIds were reversed in the
resulting ChannelInfo objects.