Revert "GitHub Actions: overwrite prior cache instead of rotating"
This reverts commit 1d4074607fe90b3183bdc287151cf11ea5a8d94a.

My understanding of the caches was incorrect.  The caches are immutable,
see https://github.com/actions/cache/blob/main/workarounds.md#update-a-cache

To keep the ccache up to date, restore the previous behavior of using a
unique key for each run.  This is the accepted method to update a cache.