Live TV select channel from another video source
In Live TV mode it is possible to select any other channel by entering the channel number.
In fixes/32 and in master this is broken; it is now only possible to select a channel in
the same video source as the current channel.
Fixed by removing a static_cast in two lines in function get_chanid that was introduced
in  commit 976989051502f54abb7173f9e9127f25f8fd6c13 on September 29, 2020.